requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pyaml==24.9.0
PyJWT==2.9.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
